move session to feature/share/func
