+++ apps/web/app/features/share/func/session.ts
