+++ apps/web/app/routes/reservation.store.tsx
@@ -5,7 +5,7 @@ import { Button } from "~/components/button";
 import { Input } from "~/components/input";
 import { storeReservationInfoSchema } from "~/features/reservation/types/schema";
 import { ProcessExplanation } from "~/features/reservation/ui/processExplanation";
-import { commitSession, getSession } from "~/session";
+import { commitSession, getSession } from "~/features/share/func/session";
 
 export const meta: MetaFunction = () => {
   return [
